api 설정 - 고

+++ app/src/main/java/com/hgh/na_o_man/domain/model/auth/UserTokenModel.kt
@@ -2,6 +2,6 @@ package com.hgh.na_o_man.domain.model.auth
 
 data class UserTokenModel(
     val accessToken: String,
-    val memberId: Int,
+    val memberId: Long,
     val refreshToken: String
 )
